test: update transaction use case tests
