+++ src/test/java/com/financebot/transaction/application/adapter/in/web/TransactionControlerTest.java
@@ -1,0 +1,4 @@
+package com.financebot.transaction.application.adapter.in.web;
+
+public class TransactionControlerTest {
+}
